Fixed defect with hard reboot not waiting long enough. Some optimizations, fixed flush.txt delimiter, handled failing fit.

+++ Esp32Mock/ESP.cpp
@@ -72,7 +72,7 @@ void shiftMicros(long long shift) {
     microShift = shift;
 }
 
-void delayMicroseconds(int delay) {
+void delayMicroseconds(unsigned long delay) {
     if (RealTimeOn) {
         microShift += delay;
     }
