+++ Esp32Mock/ESP.h
@@ -110,7 +110,7 @@ uint8_t getPinMode(uint8_t pin);
 
 void configTime(int i, int i1, const char* str, const char* text);
 void delay(int delay);
-void delayMicroseconds(int delay);
+void delayMicroseconds(unsigned long delay);
 unsigned long micros();
 unsigned long millis();
